refactor: rename hook useSnackBar to useSnackbar
changed the hook import path
