+++ src/shared/common/hook/useSnackbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./useSnackbar";
