+++ src/shared/common/hook/useSnackbar/types/index.ts
@@ -1,0 +1,2 @@
+export * as default from "./useSnackbar.types";
+export * from "./useSnackbar.types";
