+++ src/shared/common/hook/useSnackbar/types/useSnackbar.types.ts
