+++ src/shared/components/Layout/SnackBar/types/SnackBar.component.ts
@@ -1,4 +1,4 @@
-import { ISnackbarProps } from "shared/common/hook/useSnackBar/types/useSnackBar.types";
+import { ISnackbarProps } from "shared/common/hook/useSnackbar/types";
 import { FunctionReturnsVoid } from "shared/common/types/AppTypes";
 
 export interface ISnackBarView {
